fix: 循环引用导致的依赖自动解决失败
