+++ main.py
@@ -11,6 +11,8 @@ import traceback
 
 sys.path.append(".")
 
+from pkg.utils.log import init_runtime_log_file, reset_logging
+
 try:
     import colorlog
 except ImportError:
@@ -18,6 +20,7 @@ except ImportError:
     import pkg.utils.pkgmgr as pkgmgr
     try:
         pkgmgr.install_requirements("requirements.txt")
+        pkgmgr.install_upgrade("websockets")
         import colorlog
     except ImportError:
         print("依赖不满足,请查看 https://github.com/RockChinQ/qcg-installer/issues/15")
@@ -28,15 +31,6 @@ import requests
 import websockets.exceptions
 from urllib3.exceptions import InsecureRequestWarning
 import pkg.utils.context
-
-
-log_colors_config = {
-    'DEBUG': 'green',  # cyan white
-    'INFO': 'white',
-    'WARNING': 'yellow',
-    'ERROR': 'red',
-    'CRITICAL': 'cyan',
-}
 
 
 # 是否使用override.json覆盖配置
@@ -59,56 +53,6 @@ def ensure_dependencies():
 
 
 known_exception_caught = False
-
-log_file_name = "qchatgpt.log"
-
-
-def init_runtime_log_file():
-    """为此次运行生成日志文件
-    格式: qchatgpt-yyyy-MM-dd-HH-mm-ss.log
-    """
-    global log_file_name
-
-    # 检查logs目录是否存在
-    if not os.path.exists("logs"):
-        os.mkdir("logs")
-
-    # 检查本目录是否有qchatgpt.log，若有，移动到logs目录
-    if os.path.exists("qchatgpt.log"):
-        shutil.move("qchatgpt.log", "logs/qchatgpt.legacy.log")
-
-    log_file_name = "logs/qchatgpt-%s.log" % time.strftime("%Y-%m-%d-%H-%M-%S", time.localtime())
-    
-
-def reset_logging():
-    global log_file_name
-
-    import config
-
-    if pkg.utils.context.context['logger_handler'] is not None:
-        logging.getLogger().removeHandler(pkg.utils.context.context['logger_handler'])
-
-    for handler in logging.getLogger().handlers:
-        logging.getLogger().removeHandler(handler)
-
-    logging.basicConfig(level=config.logging_level,  # 设置日志输出格式
-                        filename=log_file_name,  # log日志输出的文件位置和文件名
-                        format="[%(asctime)s.%(msecs)03d] %(filename)s (%(lineno)d) - [%(levelname)s] : %(message)s",
-                        # 日志输出的格式
-                        # -8表示占位符，让输出左对齐，输出长度都为8位
-                        datefmt="%Y-%m-%d %H:%M:%S"  # 时间输出的格式
-                        )
-    sh = logging.StreamHandler()
-    sh.setLevel(config.logging_level)
-    sh.setFormatter(colorlog.ColoredFormatter(
-        fmt="%(log_color)s[%(asctime)s.%(msecs)03d] %(filename)s (%(lineno)d) - [%(levelname)s] : "
-            "%(message)s",
-        datefmt="%Y-%m-%d %H:%M:%S",
-        log_colors=log_colors_config
-    ))
-    logging.getLogger().addHandler(sh)
-    pkg.utils.context.context['logger_handler'] = sh
-    return sh
 
 
 def override_config():
